framework is done

+++ database/migrations/2018_12_03_134814_create_levels_table.php
@@ -17,9 +17,9 @@ class CreateLevelsTable extends Migration
             $table->increments('id');
             $table->string('first_lang_level_name');
             $table->string('second_lang_level_name');
-            $table->integer('self_weight');
-            $table->integer('peer_weight');
-            $table->integer('mentor_weight');
+            $table->double('self_weight');
+            $table->double('peer_weight');
+            $table->double('mentor_weight');
             $table->timestamps();
           });
     }
